requirements.txt
absl-py==0.12.0
anyio==3.6.1
apiclient==1.0.4
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
bleach==4.1.0
blis==0.7.5
CacheControl==0.12.11
cached-property==1.5.2
cachetools==4.2.2
cachy==0.3.0
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
clang==5.0
cleo==0.8.1
click==7.1.2
clikit==0.6.2
colorama==0.4.4
colorclass==2.2.2
contextvars==2.4
crashtest==0.3.1
cycler==0.10.0
cymem==2.0.6
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.4
docopt==0.6.2
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.4.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-api-core==1.27.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h11==0.12.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.13.7
httplib2==0.19.1
httpx==0.19.0
hyperframe==6.0.1
idna==2.10
immutables==0.16
importlib-metadata==3.10.1
importlib-resources==4.1.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.6.0
keras-lr-finder==0.1
keras-models==0.0.7
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==22.3.0
kiwisolver==1.3.1
langcodes==3.2.1
librosa==0.8.0
llvmlite==0.36.0
lockfile==0.12.2
logging==0.4.9.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
msgpack==1.0.4
murmurhash==1.0.6
musdb==0.4.0
museval==0.4.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
norbert==0.2.1
notebook==6.4.10
numba==0.53.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.4.58
openunmix==1.2.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pastel==0.2.1
pathlib==1.0.1
pathy==0.6.1
patsy==0.5.2
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.8.2
platformdirs==2.4.0
playsound==1.2.2
poetry==1.1.13
poetry-core==1.0.8
pooch==1.3.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.17.0
ptyprocess==0.7.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyDrive==1.3.1
pydub==0.25.1
Pygments==2.12.0
pylev==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pysndfx==0.3.6
python-dateutil==2.8.1
pytz==2021.1
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==1.1.6
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==0.24.2
scikit-posthocs==0.6.7
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
shellingham==1.4.0
simplejson==3.17.6
six==1.15.0
smart-open==5.2.1
sniffio==1.2.0
SoundFile==0.10.3.post1
sox==1.4.1
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spleeter==2.3.1
srsly==2.4.2
statsmodels==0.12.2
stempeg==0.2.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.12.1
terminaltables==3.1.10
testpath==0.6.0
thinc==8.0.13
threadpoolctl==2.1.0
tomlkit==0.11.0
torch==1.10.2
torchaudio==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typer==0.3.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.14.1
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.0
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
